requirements.txt
appdirs==1.4.3
bonobo==0.6.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.30
Django==2.2.7
django-background-task==0.1.8
django-background-tasks==1.2.0
django-compat==1.0.15
django-rest==0.0.1
djangorestframework==3.10.3
fs==2.4.11
graphviz==0.8.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mondrian==0.8.0
mysqlclient==1.4.6
packaging==19.2
pbr==5.4.4
pretty-cron==1.2.0
psutil==5.6.7
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
python-slugify==1.2.6
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
stevedore==1.31.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.7
whistle==1.0.1
